requirements.txt
# Aktualizator Strony - prakt.dziadu.dev
# Wymagania - v5.0

# Główna aplikacja
# Python 3.7+ - wbudowany Tkinter

# Zależności (v4.1-v5.0)
beautifulsoup4>=4.9.0      # Parsowanie HTML
python-dotenv>=0.19.0      # Environment variables (.env support)
customtkinter>=5.0.0       # Modern Dark/Light Mode GUI
darkdetect>=0.8.0          # Automatyczne wykrywanie dark mode

# v4.2 Features - Analytics & Reports
openpyxl>=3.0.0            # Eksport do Excel
reportlab>=3.6.0           # Eksport do PDF

# v5.0 Features - Integracje & Scheduling
schedule>=1.1.0            # Harmonogram aktualizacji (Scheduler)
requests>=2.28.0           # HTTP requests (Webhooks, Slack, Discord)
PyGithub>=1.55.0           # GitHub API (Webhooks)
slack-sdk>=3.19.0          # Slack Integration
discord.py>=2.0.0          # Discord Integration
email-validator>=1.3.0     # Email validation

# v5.0 Features - Web Dashboard
flask>=2.3.0               # Web framework dla dashboard
flask-cors>=4.0.0          # CORS support
flask-socketio>=5.3.0      # WebSocket support

# v5.0 Features - Database
sqlalchemy>=2.0.0          # ORM dla SQLite
alembic>=1.11.0            # Database migrations

# v5.0 Features - Utilities
cryptography>=40.0.0       # SSH Keys encryption
paramiko>=3.0.0            # SSH support
GitPython>=3.1.0           # Python Git wrapper
pyinstaller>=5.0.0         # Executable build

# Narzędzia dla developers
pytest>=6.0.0              # Unit testy
pytest-cov>=2.10.0         # Coverage raport
black>=23.0.0              # Code formatter
pylint>=2.17.0             # Code linter

# Biblioteki standardowe (wbudowane w Python):
# - tkinter (GUI)
# - subprocess (obsługa Git)
# - pathlib (zarządzanie ścieżkami)
# - json (konfiguracja)
# - threading (wątki)
# - datetime (data/czas)
# - re (wyrażenia regularne)
# - logging (logowanie)
# - shutil (operacje na plikach)

# Aby zainstalować zależności:
# pip install -r requirements.txt

# Aby uruchomić testy:
# pytest tests/ -v

# Aby uruchomić:
# python apk.py

